Sync data in background triggered by broadcast
This technique will not work with persistence enabled due to the single use of the listener,
and how internally Firebase triggers a value update for cached data, followed by a
second update for online data.

Workarounds include writing dummy data to the database before fetching the new values,
or adding a timer to kill the listener after a certain period instead of a single value.